RNDSVC-2925: Rebase ipxe from upstream (v1.21.1-411-f4193533)
X-Arc-Original-Commit: ccb9bc3b9a75361c50ddb3c853a3ffb729752061

+++ src/drivers/usb/usbkbd.c
@@ -95,7 +95,7 @@ static unsigned int usbkbd_map ( unsigned int keycode, unsigned int modifiers,
 		}
 	} else if ( keycode <= USBKBD_KEY_UP ) {
 		/* Special keys */
-		static const uint16_t special[] = {
+		static const unsigned int special[] = {
 			0, 0, 0, 0, 0, KEY_F5, KEY_F6, KEY_F7, KEY_F8, KEY_F9,
 			KEY_F10, KEY_F11, KEY_F12, 0, 0, 0, KEY_IC, KEY_HOME,
 			KEY_PPAGE, KEY_DC, KEY_END, KEY_NPAGE, KEY_RIGHT,
@@ -110,7 +110,7 @@ static unsigned int usbkbd_map ( unsigned int keycode, unsigned int modifiers,
 		if ( leds & USBKBD_LED_NUM_LOCK ) {
 			key = "1234567890." [ keycode - USBKBD_KEY_PAD_1 ];
 		} else {
-			static const uint16_t keypad[] = {
+			static const unsigned int keypad[] = {
 				KEY_END, KEY_DOWN, KEY_NPAGE, KEY_LEFT, 0,
 				KEY_RIGHT, KEY_HOME, KEY_UP, KEY_PPAGE,
 				KEY_IC, KEY_DC
